more precise fun-decl, less precise cxx-block
Having :cxx-stmt in the cxx-block definition was inducing some really
pathological behavior in some (but not all) deep trees. Not entirely
clear why this was the case, but it’s better now.

Expander has also received tooling that can be used for timing various
operations in the expansion code, since the profiler has trouble with
it.